refactor to hittop directory structure
