+++ hittop/parser/concat-test.cc
@@ -1,4 +1,4 @@
-#include "casper/concat.h"
+#include "hittop/parser/concat.h"
 
 #include <string>
