+++ hittop/parser/concat.h
